Converted the time scales used for x sigmas to length scales.
There is clearly as issue with the SEM sigma spec for the x sigmas.
So we are converting the time scales into length scales just by
multiplying by the bulk velocity. May be more accurate to multiply by the
local velocity, but im not sure. This significantly reduces the required
number of eddies for the compressible cases.