True local convection speeds
After some thought, its obvious that we can't convect through
the domain at a single speed. Bc the eddys above and below
convect at slightly different speeds. So we now account
for this with an x_offset. Also added a warning if any
points sent to generate_primes lie completely outside the domain.