Fixed the issue with local convection speed
The eddy shape function is unbounded for performance, and it assumes
we have already filtered out the eddys that are more that a sigma away
from the point of interest.

The local convection speed x_offset was not being applied to the second x_dist
calculation, the the shape function was getting massive values.