Yikes... cant interpolate in space or time.
Turns out that interpolating is bad for the desired statistics. So
we no longer interpolate in space. We also now have an extra
input into sem.inp that determines if we are interpolaing in time.
If we are, we approximate the unnormalized flutuations as a continuous
signal and then normalize that signal, and then downsample the values
at the frame locations and export that. This gives good results
regardless of the approximation-to-interpolation between frames difference.

Currently we just hard code the continuous approximation with 10 points between
frames.